+++ tests/test_coding/test_transforms.py
@@ -3,6 +3,7 @@ import operator
 import unittest
 import base64
 
+from satella.coding.predicates import x
 from satella.coding.transforms import stringify, split_shuffle_and_join, one_tuple, \
     merge_series, pad_to_multiple_of_length, clip, b64encode, linear_interpolate, \
     hashables_to_int, none_if_false, merge_list, is_subset, unpack_dict
@@ -13,6 +14,10 @@ class TestTransforms(unittest.TestCase):
     def test_unpack_dict(self):
         a, b, c = unpack_dict({1: 2, 2: 3, 4: 5}, 1, 2, 4)
         self.assertTrue(a == 2 and b == 3 and c == 5)
+
+        a, b, c = unpack_dict({1: 2, 2: 3, 4: 5}, 1, 2, 4,
+                              map_through=x*2)
+        self.assertTrue(a == 4 and b == 6 and c == 10)
 
     def test_is_subset(self):
         self.assertTrue(is_subset({}, {}))