+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 # v2.18.4
+
+* added `map_through` to `unpack_dict`